:art: Added AtomicTrigger abstract class.
AtomicTrigger is an abstract class defined for the concept "Button" and "Switch". Button can be interpreted in many ways: A pressure button, or a player interact-able button using F key. This abstract class only defines two ways of Enabling, Disabling &amp; Triggering the target react machine. The detailed condition of when or how to do this is described in its inheritors.