:sparkles: Added MergeSignalGate.
MergeSignalGate (MSG) is a gate that process multiple sources of "enable" signals.
MSG takes multiple inputs of triggers as a react machine, and gives an output as an atomic trigger.
That is to say, MSG is a ReactMachine, and also an AtomicTrigger.
However, it inherits the class ReactMachine, since it needs most of the methods from ReactMachines.